refactor: 입력값 검증 부분 BaseBallInputValidator로 분리.
